+++ src/main/java/com/graphql/deliveryShare2/sample/AboutUser/UserRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
-    UserEntity findBySeq(int seq);
+    UserEntity findBySeq(Long seq);
 }
